fix(regression): Undo most changes from 9ddafd03501c8688ccd579585585bf9725c2ebf3
The change made in block.cpp actually broke LOOM-EGA-FR, see issue #21.

So, drop it, and drop the change in scummfont.cpp as well, just in case.

Keeping the file.hpp changes though, because they should be safe, since
that's how bswap16() and bswap32() are often implemented (but we'll
revert if if it's known to cause other bugs or regressions).