feat(scummtr): Keep the two "quote" symbols different, even when -c is used
By default, SCUMM games often render the quotation mark symbol (0022)
and the grave accent (0060) with the same glyph, but for some cases or
for some languages, it's probably better to output them as different
characters, as is done is the default ASCII mode (i.e. when -c is not
used).