feat(scummtr): Recognize even more CP 850 characters
It appears that most SCUMM games actually follow the MS-DOS CP 850
code page internally, so we can try recognize any "useful" but missing
CP 850 character, as long as they're available at the same codepoints
in both Windows-1252 and ISO-8859-1, and as long as no SCUMM game
already uses that codepoint.

Warning: recognizing it doesn't mean that it will work in-game. It
may not have been drawn in the included font; use scummfont for that,
if necessary. The idea is to settle on a safe subset/variant of the
MS-DOS CP 850 code page.

This should improve compatibility for Western European alphabets such
as the ones used in the French, Italian, Spanish, Irish, Portuguese,
Welsh, Icelandic, or Scandinavian languages. They're also a bit of
improvement for some Eastern European alphabets.

It is still incomplete and biased towards Western European languages.
Sorry for that, but this is the historical behaviour of these games.
Hopefully we'll find another way of improving the situation for other
languages, alphabets and users… in the meantime, it's still possible
to use the ASCII mode, and do your local modifications to the included
fonts, but it's not convenient.