feat(releases): Make it explicit that we use an older Mingw-w64 version on purpose
Mingw-w64 developers recommend using an older version if old Windows
versions are targeted. Windows XP is old, and newer versions of Mingw-w64
seem to focus on Windows 10 by default.

So, we're keeping the older version from Debian 9 on purpose, and we
have no reason to upgrade for the moment (our tools are relatively
simple, and Windows XP support still makes sense for old DOS
games).