feat(scummtr): Add a new `-r` flag to preserve the original text encoding
By default, non-ASCII characters are handled as escape sequences, but this
is not very practical.  We also have a `-c` flag, but this only improves the
situation for Western-European languages.

This new `-r` flag just interprets the original text "as-is", with no
transformation.  This improves the situation for some languages such
as Japanese, although you still need a text editor which recognizes the
original encoding (often an MS-DOS code page).

(Also rename the internal _escaped to !_binary, while there.)