feat(scummtr): Recommend "-A aov" by default, and explain it better
Letting the tool do a lot of extra resource padding by default is just
way easier in most cases, and the increased memory usage has no impact
on most machines released in the last 30 years.

This should make sure, in most cases, that most translations remain
compatible with the original interpreters, with no extra work required
(ScummVM appears to already include some extra protections for this).

Of course, you're still free to omit this option if you have your
reasons for that.