feat: Reject some versions of MONKEY1-FLOPPY-VGA, because of a corruption bug
Some versions of MONKEY1-FLOPPY-VGA (possibly only the earliest 1.0
version) have a duplicate 117/10 costume inside room 59 (Stan's).

Commit dc94b2a5bafdeba50f4e59dd91278610899edd8a tried to ignore the
duplicate, as is already done for duplicate scripts, but it looks like
this doesn't work as expected for duplicate costumes, because ScumMVM
just glitches when displaying the modified game (as 000.LFL appears
to contain some invalid bytes).

I couldn't find an easy fix for this, and it's probably too late to
spend more time on this for the 0.5.0 release, so let's just reject
this game variant for now (earlier releases didn't support it at all).

Issue #47.