feat(scummfont): Reject any BMP file with a modified palette
It seems that ScummFont just expects that its palette to remain
exactly the same between output and input usages, but most modern
image editors and tools tend to change/reorder it, so ScummFont
can't reliably determine if a pixel has some content or not, thus
breaking (possibly among other things) its detection of character
widths, for example with Indy3 or Loom, if one edits its BMP file
with Apple Preview.app and then ppmtobmp (issue #15).

For this reason, rejecting any BMP file whose palette has been
modified just feels safer, at the moment.  The problem is that the
only "reliable" BMP editor for ScummFont seems to be MS Paint, on
Windows XP and older... but fixing this properly would require
quite a lot of code rewrite (the BMP format appears to be simple
but it's not really reliable to use your own encoder/decoder for
it), and we need to make a release, so better safe than sorry.

As an extra precaution, commit c80e6667bfd438c53c6451dea7a4fca2a92241af
has also been reverted.