feat(scummfont): Recommend the GIMP image editor for modifications
BuXY reported that the GIMP does appear to preserve the original BMP
settings (and the palette), as long as you enable the "Do not write
color space information" compatibility option, when exporting to BMP.

So, recommend this, especially since GIMP is free and available on
many systems.  It's probably easier to work around the ScummFont BMP
limitations with GIMP than with an old MS Paint release.