feat(releases): Move the MS-DOS builder to the djgpp-toolchain Ubuntu PPA
It appears to be more maintained than the project we were using until
now, and the very nice bonus is that its Ubuntu 19.04 packages were
also available for arm64, which means that it's now possible to do a
native Docker arm64 build instead of relying on the really slow x86
emulation from Docker.

Ubuntu 19.04 itself is EOLd but that's not a problem for our usage.