feat(releases): Move the Win32 builder from Debian 10 to Ubuntu 19.04
The MS-DOS builder currently imposes Ubuntu 19.04 (that's the only
release where there is an arm64 DJGPP PPA, currently), so let's reuse
this in order to avoid having to download 3 different Linux images.

Fortunately, Ubuntu 19.04 is very close to Debian 10 for what we're
doing here.  The only downside is that it's now unmaintained, but it
doesn't really matter for a small builder image.