fix(releases): Revert the last two Ubuntu 19.04 commits
The djgpp-toolchain Ubuntu PPA is not reliable enough.  On arm64, ranlib
had to be run to fix some indexes; on amd64 that's not required but it
fails linking with any C++ static library instead.  Updating to the last
release on Ubuntu 20.04 doesn't fix the problem either.

Go back to the older toolchain which is amd64-only but is OK with
static libraries.

This reverts commit b5d6e1132608c4fbd7289d47b45be1ab537647ec.
This reverts commit 4d6b1d580df052a7a0f0ecb41f95b5a02114433a.