feat(scummtr): Fix text extraction resource for German releases of Indy3 (#58)
This is an original bug in the official German releases, some German Eszett
characters were misencoded as \255 instead of \225, triggering the "special
espace sequence" code.

The following script lines are impacted:

[023:SC#0108]^irgendwo hier im Schlo\255.
[028:SC#0071]Nicht so schnell!  Woher weiß ich, da\255 Sie kein Spion sind?

Just ignore the \255 and replace it with the inteded \225, on the fly.
This unbreaks exporting/importing the text with ScummTR.

See also the related original ScummVM bugs:

    https://bugs.scummvm.org/ticket/1675
    https://bugs.scummvm.org/ticket/2715

Thanks to @gabberhead for reporting the issue.