feat(common): Improve _xorBuffer() performance
Thomas' original code had some logic to xor the values with 4 bytes
at once, to improve performance.

However, this caused -Wcast-align issues on some platforms (see issue
ecff30b86c022aa60941a8f71f7e7bc26e96f492.

This new approach restores a bit of this logic, but it's now done in
a way to avoid unaligned loads. It seems to work on that old mips64el
board I have.

Since it appears that _xorBuffer() is called on a majority of very
small buffers, it doesn't make sense to set up this "pipeline" if
there are just a few bytes to copy, though.