feat(build): Silence some -Wformat cases (not fixing them) to enable the flag again
Silence the non-trivial ones regarding file sizes. This makes use of
pragmas, so we now need to ignore unknown pragmas.

The pragmas need to be kept outside of the functions, for older
GCC releases and such. For GCC < 4.6, it's hopeless; the current
warnings will always be printed.  We have to fix them at some point,
anyway.