feat(scummtr): Improve handling of CRLF files used without `-w`
Several fan translations exist with bogus `\013` characters at the
end of each string.  This can cause subtle issues with ScummVM or
with the original interpreters, since it's not really expected.

Such characters would come from doing an initial text export with
the `-w` option, but forgetting to add this option again in all
subsequent import calls.  ScummTR would then inject the character
as is into the game resources.

Since there are several instances of fan-made translations having
this issue (and since the existing UX isn't that great), it seems
reasonnable to warn users about such cases.

That is: if the game hasn't been "corrupted" with bogus `\013`
yet, try to be helpful, and add `-w` if we'd need it but it's
been forgotten.  If the damage has already been done, report
this as a fatal error, and tell the user that a manual fix is
required (it feels safer to do so).