feat(scummtr): Warn when the "Indy3 Mac hack" is used elsewhere
I was already aware that this warning was also displayed for (all? most?)
EGA titles, even though the original comment only mentions Indy3 Mac.

But now I see that this hack might be related to issue #60 (where
disabling this hack appears to "fix" part of this issue).

So it feels safer to print a warning, when this hack is triggered in
other places -- at least so that there's a better understanding of
when it's applied.