fix(win32): Fix win32 path cleanup loop
strchr() would always be called on params[k].value, which is a bit
inefficient.

The other strchr() call could be changed to `p = strchr(p, '\\')`,
but let's just use a simpler loop; it's going to parse relatively
small strings anyway, so frequent calls to strchr() are not going
to be worth it.