feat(scummtr): Warn if one uses `-w` but there's no CRLF on that line
Try to be useful and fix that on the fly for users, but still, print
a warning for now.

I don't know what's the best approach: give a single warning, warn on
each faulty line (can be big; current default), fix the error silently
or make it a fatal error.